# Improve performance of inverse trigonometric functions