<!-- Your title above should be a short description of what
was changed. Do not include the issue number in the title. -->

#### References to other Issues or PRs
<!-- If this pull request fixes an issue, write "Fixes #NNNN" in that exact
format, e.g. "Fixes #1234" (see
https://tinyurl.com/auto-closing for more information). Also, please
write a comment on that issue linking back to this pull request once it is
open. -->


#### Brief description of what is fixed or changed

This PR improves performance of the inverse trigonometric functions by:

* Caching the lookup tables in `InverseTrigonometricFunction` (so the tables are not constructed for each invocation of the function)
* Consistently using the variable `pi` instead of `S.Pi` (which saves an attribute lookup)

The main speed improvement is for numbers. Two benchmarks:

```
%timeit atan(np.random.rand()) 
#main: 670 µs ± 19.4 µs per loop (mean ± std. dev. of 7 runs, 1000 loops each)
#PR: 488 µs ± 1.47 µs per loop (mean ± std. dev. of 7 runs, 1000 loops each)
```

```
%timeit asin(np.random.rand()) # 670 us on main, 512 us with class init of _atsin_table
#main: 865 µs ± 10.9 µs per loop (mean ± std. dev. of 7 runs, 1000 loops each)
#PR: 513 µs ± 705 ns per loop (mean ± std. dev. of 7 runs, 1000 loops each)
```
#### Other comments

The same improvements can be applied to other functions in the sympy codebase. The changes in this PR are limited to a single file to keep the PR small and first discuss whether the approach is good.

For catching the sympy `cacheit` method is used. Since the lookup table methods have no arguments, this is a bit overkill. We could replace it by `functools.lru_cache(size=1)`

Another approach would be to replace the `_atan_table` with a class attribute. This failed because of circular references.

#### Release Notes

<!-- Write the release notes for this release below between the BEGIN and END
statements. The basic format is a bulleted list with the name of the subpackage
and the release note for this PR. For example:

* solvers
  * Added a new solver for logarithmic equations.

* functions
  * Fixed a bug with log of integers.

or if no release note(s) should be included use:

NO ENTRY

See https://github.com/sympy/sympy/wiki/Writing-Release-Notes for more
information on how to write release notes. The bot will check your release
notes automatically to see if they are formatted correctly. -->

<!-- BEGIN RELEASE NOTES -->
* functions
  * Improve performance of trigonometric functions
<!-- END RELEASE NOTES -->