@oscargus 2022-05-29
> is for numbers

I guess that there would be an even bigger improvement if the argument was checked for being a `Float` first and in that case skip all the remaining tests? This probably holds for most evaluations as they focus on symbolic values.