@eendebakpt 2022-05-29
@oscargus I tested skipping the tests in `atan.eval` with a check `isinstance(arg, Float)`. Most tests don't have a significant impact on the running time. The only exception is `i_coeff = arg.as_coefficient(S.ImaginaryUnit)`, which is quite expensive.

```
%timeit atan(np.random.rand()) 
#main: 670 µs ± 19.4 µs per loop (mean ± std. dev. of 7 runs, 1000 loops each)
#PR: 488 µs ± 1.47 µs per loop (mean ± std. dev. of 7 runs, 1000 loops each)
#PR + fix as_coefficient: 211 µs ± 882 ns per loop (mean ± std. dev. of 7 runs, 1000 loops each)
```
I consider two options to fix the `rg.as_coefficient(S.ImaginaryUnit)`:

1) Add a check with `isinstance(arg, Float)` in the `eval` method. Works, but has to be done at multiple locations (a quick grep on `as_coefficient(S.ImaginaryUnit)` gives 36 hits
2) Improve `arg.as_coefficient`. Something like
```
    def as_coefficient(self, expr):
        if isinstance(self, Float) and expr == S.ImaginaryUnit:
            return None
        
        r = self.extract_multiplicatively(expr)
        if r and not r.has(expr):
            return r
```
works, but it not really generic.

I would also be fine with creating a new issue for this to keep this PR focused.