@eendebakpt 2022-05-30
@oscargus If all expressions of interest have `.is_real`, we could replace all occurrences of `as_coefficient(S.ImaginaryUnit)` with a specialized method:
```
def as_coefficient_imaginary_unit(arg):
   if arg.is_real:
      return None
   else:
      return arg.as_coefficient(S.ImaginaryUnit)
```