@oscarbenjamin 2022-05-31
Using `is_real` is a semantic check whereas I think what is really needed here is a structural one (also for performance reasons because `is_real` can be slow).

Consider something like:
```python
In [10]: x = symbols('x', imaginary=True)

In [11]: e = x*I

In [12]: e
Out[12]: ⅈ⋅x

In [13]: e.is_real
Out[13]: True

In [14]: sin(e)
Out[14]: ⅈ⋅sinh(x)
```
This PR changes that behaviour:
```python
In [3]: sin(e)
Out[3]: sin(ⅈ⋅x)
```