@oscarbenjamin 2022-05-31
> That is interesting, but probably for another PR.

I disagree. If it's better to do it there then there's no need to go around making changes to each of the subclasses. Also I suspect that skipping eval altogether would give a bigger speedup than making minor changes in eval.