@eendebakpt 2022-05-31
> > That is interesting, but probably for another PR.
> 
> I disagree. If it's better to do it there then there's no need to go around making changes to each of the subclasses. Also I suspect that skipping eval altogether would give a bigger speedup than making minor changes in eval.

@oscarbenjamin This PR contains 3 optimizations:

* Cache the trigoniometric tables
* Replace `S.Pi` with `pi`
* Fast path for `Float` arguments via `_imaginary_unit_as_coefficient`

The first two seem worthwhile because they are also effective for non `Float` arguments. The last one is indeed irrelevant if we can implement your suggestion to modify the parent class.

I created PR #23565 to test the changes to the parent class. I will put this PR on hold until that PR is either merged or rejected.