@eendebakpt 2022-06-08
> > For catching the sympy cacheit method is used. Since the lookup table methods have no arguments, this is a bit overkill. We could replace it by functools.lru_cache(size=1)
> 
> If cacheit is slow for this case (which we should check if it actually is), then we should optimize this at the cacheit level, so that it just saves the result for functions without arguments rather than using a lru_cache. However, I suspect this optimization may already exist in lru_cache, so it's possible this doesn't even matter.

The overhead of cacheit is not a lot, so I think this is fine. (the overhead is in checking for unhashable types, which would be hard to change). A special check for functions without arguments is possible, but not worth the effort here.