@eendebakpt 2022-06-10
@oscarbenjamin As #23582 is not a good alternative for the `_imaginary_unit_as_coefficient`, can you review again?