@oscarbenjamin 2022-06-12
Timing here I don't see much of a difference between the PR and master.

This is master:
```python
In [1]: %timeit atan(np.random.rand())
954 µs ± 4.98 µs per loop (mean ± std. dev. of 7 runs, 1000 loops each)

In [2]: %timeit asin(np.random.rand())
1.43 ms ± 15.1 µs per loop (mean ± std. dev. of 7 runs, 1000 loops each)
```
This is the PR:
```python
In [1]: %timeit atan(np.random.rand())
920 µs ± 4.35 µs per loop (mean ± std. dev. of 7 runs, 1000 loops each)

In [2]: %timeit asin(np.random.rand())
1.59 ms ± 120 µs per loop (mean ± std. dev. of 7 runs, 1000 loops each)
```
What happened to the speedups reported in the OP?