@oscarbenjamin 2022-06-13
I'm not sure what happened there. I just tried (possibly on a different computer because I don't remember which I used before) and I can see the difference.

This is master:
```python
In [1]: %timeit atan(np.random.rand())
911 µs ± 58.7 µs per loop (mean ± std. dev. of 7 runs, 1000 loops each)

In [2]: %timeit asin(np.random.rand())
1.39 ms ± 4.96 µs per loop (mean ± std. dev. of 7 runs, 1000 loops each)
```
This is the PR:
```python
In [1]: %timeit atan(np.random.rand())
326 µs ± 7.4 µs per loop (mean ± std. dev. of 7 runs, 1000 loops each)

In [2]: %timeit asin(np.random.rand())
523 µs ± 1.19 µs per loop (mean ± std. dev. of 7 runs, 1000 loops each)
```